Update README.md
summary:-
Passionate about coding and eager to contribute to innovative projects, I am an aspiring Python Developer with proficiency in MySQL, CSS, MongoDB, and Django. With a strong foundation in web development and a keen eye for detail, I thrive in dynamic environments where problem-solving and continuous learning are paramount. My goal is to leverage my technical skills to develop efficient and user-friendly solutions, while continually expanding my knowledge in the tech field. Let's connect and explore opportunities to collaborate on impactful projects